Move faction data retrieval to lib
